cloud_storage/inventory: Add high level API
The high level API which will eventually be used by the scrubber service
is added here. It stores the vendor specific API as a variant field, and
simply visits the variant for any operations.

The low level API now inherits from a base_ops interface. The purpose is
not runtime polymorphism but rather enforcing constraints on the variant
stored in the high level API, such that any type added to the variant
must provide methods from base_ops.