rpk: Add raid0 subcommand
This command creates, formats and mounts a Linux RAID0
device, intended to be used as data directory with Redpanda.

The RAID0 is optimized by default for cloud environments with ephemeral SSD disks.
Any data in the RAID member devices will be deleted, all blocks will discarded and
superblocks erased.

The setup includes adding the RAID device definition into mdadm.conf(5) and
updating initramfs so it is re-assembled correctly on boot, either in the same
machine or a different machine (in case of host migrations).

The RAID0 device is formatted with XFS and mounted using a systemd mount unit
placed under /etc/systemd/system and enabled to mount the RAID device on boot.