cst/inv: Add inventory-config-exists helper
A helper is added to check if an inventor configuration exists already.
This is expected to be used as a check before creating the inventory
configuration.

While the GET call used for the check returns the inventory config
details, we do not use these other than the HTTP status code. For the
AWS implementation added here the details are not needed. For GCP we
will need to extract the server assigned UUID for future calls.