cst/inv: Add maybe-create inventory config API
A top level method is added to conditionally create inventory config if
it does not already exist. This involves two operations, one download
(to check if the config exists) and one upload (to create). The results
of the two are merged into a success/failure enum class.

This top level method is expected to be used by callers to create the
config.