storage: Use fragmented_vector in OT-state
offset_translator_state uses vector to store entries. This may create
oversized allocations when the struct is serialized/deserialized. This
commit replaces vector with chunked_vector.

Both serialized data structure and serialization code path are updated.
The version of the data structure is not changed because serde treats
any vector-like structure the same and uses the same format.