cloud_storage: skip over segments below archive start
This fixes a bug for timequeries that could result in a removed segment being
unsuccessfully downloaded during a timequery. The sequence of events is as
follows:

1. Spillover manifest [7278, 8592]
2. Spillover more manifests... [...13574]
3. Eventually our stm manifest is [13875, 15802]
4. We housekeep and truncate to 7804. This is the new archive start offset and
   archive clean offset
5. A timequery comes in whose timestamp falls in manifest [7278, 8592]
6. Segment [7278-7410] is attempted to be downloaded, which has already been
   removed

To fix this, as we're seeking within a given manifest, we'll now check against
the overall start offset of the partition, skipping any segments that fall
below the start.

Fixes #15042