cloud_storage/topic_manifest: introduce topic_properties_mapping
wiring is performed in the next commit

this a collection of type_mapping objects, each describing how a field
of cluster::topic_properties is mapped to json.

type_mapping is a struct that defines a mapping for a cluster::topic_property field and json.
since the fields are optional/tristate of basic types, field_mapping is not
too complicated.

The mapping consist of a c-string name, a member pointer to
the field in topic_properties, and a pair of functions to perform the
mapping of the value.

For optional<T>/tristate<T>, the to_json function will
be called only if the field has a value and it will receive the unwrapped  T
const & value.

The from_json function will be called with
std::string_view/uint64_t/int64_t json value and their result will be
assigned to the field. the argument type will determine the accepted
parameter type.

a typical use to map topic_properties::compression, that
will perform string<->enum_value mapping looks like this
  type_mapping{
    "compression",          // name
     &tp_p::compression,    // member field pointer
     boost::lexical_cast<std::string, model::compression>,          //
     to_json mapper boost::lexical_cast<model::compression,
     std::string_view>},    // from_json mapper

to support sub-object, another type is provided: aggregate_type_mapping.
This takes a c-string name and a member pointer, but defers to a separate root_type_mapping
for instructions on how to serialize a struct. root_type_mapping is a collection of
type_mapping and aggregate_type_mapping.

in the deserialization phase, each aggregate_type_mapping will create a context object
to keep track of the sub-objects that is being deserialized at any moment. context will nest
in a stack inside the handler.

an example: the following json
{
    "remote_topic_properties": {"revision_id": 2}
}
is decoded with this sequence of events

StartObject({)                 - push context pointing to topic_properties
Key("remote_topic_properties") - record "remote_topic_properties" as the current key
StartObject({)                 - find aggregate_type for "remote_topic_properties",
                               - push context pointing to topic_properties.remote_topic_properties
Key("revision_id")             - record "revision_id" as the current key
Value(2)                       - find type_aggregate for revision id,
                               - set topic_properties.remote_topic_properties.revision_id=2
EndObject(})                   - pop context
EndObject(})                   - pop context