tests: added test for virtual connections
Added a test that validates if requests executed in different connection
contexts do not block each other. The test is using a `kafka-python`
client. Unfortunately we need to access the client internal to change
client ids and expected response sequences but we are able to execute
the virtual connection handling code without pulling in MPX into
ducktape.

Signed-off-by: Michal Maslanka <michal@redpanda.com>