test: wait for cluster to become healthy after setup
Immediately after the test starts up and the initial topics are created
the cluster health is tested. In this case the cluster was unhealth
because of:

    reasons:                [leaderless_partitions under_replicated_partitions]',

But these aren't really indications of an _unhealthy_ cluster, but
rather normal states that should be considered unhealthy if they
persist. In the test that failed 100s of partitions were being created
in a single topic, so perhaps something went out for coffee temproarily.

Anyway, we can give some grace period to becoming healthy rather than
failing.

    AssertionError('cluster unhealthy before start of test')
    Traceback (most recent call last):
      File "/usr/local/lib/python3.10/dist-packages/ducktape/tests/runner_client.py", line 182, in _do_run
        self.setup_test()
      File "/usr/local/lib/python3.10/dist-packages/ducktape/tests/runner_client.py", line 260, in setup_test
        self.test.setup()
      File "/home/ubuntu/redpanda/tests/rptest/redpanda_cloud_tests/high_throughput_test.py", line 357, in setup
        super().setup()
      File "/home/ubuntu/redpanda/tests/rptest/tests/redpanda_cloud_test.py", line 41, in setup
        assert self.redpanda.cluster_healthy(
    AssertionError: cluster unhealthy before start of test

Fixes: https://github.com/redpanda-data/redpanda/issues/16670

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>