archival: Add archiver_service
The service is running on every shard. It subscribes to manage/unmanage
notifications from the 'cluster::partition_manager' and also leadership
notifications from the 'raft::group_manager'. On every shard it
maintains a set of managed partitions. These are partitions which are
created on the same shard by the 'cluster::partition_manager'. When the
service receives notifications from the 'raft::group' manager and
detects that the partition becomes a leader on a shard it creates and
start the 'ntp_archiver' instance.

The service is implemented using Boost MSM library. This is needed
because the service maintains a state-machine per NTP. The transitions
are not that complicated on their own. But there is a complication.
The notifications from both `cluster::partition_manager' and
`raft::group_manager' are not futurized (the callbacks return void). But
the start/stop of the ntp_archiver is an asynchronous operation. This
doubles number of states because instead of 'passive' and 'active' we
now have 'passive', 'active', 'starting', and 'stopping'.

It also complicates state transitions (the number of transitions
quaruples). Without async events we would have only leadership events
(acquired leadership and lost leadership). Now we have four more events:
archiver started, archiver stopped, archiver start failed, archiver
shutdown failed. We can also receive leadership events when the archiver
is starting or stopping.

This is how main state transitions look (slightly simplified):

         stopped
         archiver┌───────────┐ became leader
         ┌──────►│  passive  ├───────┐
         │       └───────────┘       │
         │                           ▼
    ┌────┴─────┐  lost          ┌──────────┐
    │ stopping │◄───────────────┤ starting │
    └──────────┘  leadership    └────┬─────┘
         ▲       ┌───────────┐       │
         └───────┤ active    │◄──────┘
         lost    └───────────┘ started
         leadership            archiver

Initially, the object is created in 'passive' state. When partition
becomes a leader it transitions into the 'starting' state. In this state
the archiver is starting asynchronously in the background. If the
leadership was lost during transition the FSM transitions to the
'stopping' state in which the archiver is stopped asynchronously in the
background. In the code it's implemented using MSM 'Defer' functor.
The defer functor is used to implement simple even queueing in the FSM.
When the even triggers Defer functor it's queued and re-triggered
after the state transition. In this case we have a row that maps
'starting' state and the `leadership_lost' event to the 'Defer' functor.
This means that if the 'leadership_lost' event will be received while
FSM is in 'starting' state it will be queued. When the FSM will
transition to the 'active' state this even will be re-triggered and the
FSM will transition to the 'stopping' state!

The same approach is taken with the 'stopping' state. If the leadership
is acquired while the FSM is in the 'stopping' state we're transitioning
to 'passive' state and then immediately to the 'starting' state via the
'Defer' functor. There are two entries in the transition table for that.

    Failed to    ┌───────────┐  Failed to create
    stop ┌──────►│  passive  │◄──────┐ or start
         │       └───────────┘       │ archiver
         │                           │
    ┌────┴─────┐                ┌────┴─────┐
    │ stopping │                │ starting │
    └──────────┘                └──────────┘

The error handling is pretty simple. If either 'stopping' or 'starting'
state fails (the bg op fails) the FSM transitions to the 'passive'
state.

The 'passive' state is the state in which we can safely dispose the FSM.
In any other state the 'ntp_archiver' exists. It's either working or
starting or stopping. During the sthutdown of the FSM we're setting the
'promise<>' object to the 'passive' state and then waiting for it.
When the 'passive' state is entered it first checks if the 'promise<>'
field is set. If this is the case it invokes 'set_value' if transition
was caused by clean shutdown of the archiver or it invokes
'set_exception' and the error is propagated further. Note that the
'gate' is not used to control the lifetime of the background operations.
This is not needed because we have the invariant that guarantees that
the FSM first returns to the 'passive' state.

Both 'stopping' and 'starting' states are holding units from the
'_archiver_reset_mutex' semaphore owned by the 'cluster::partition'.
This solves the following lifetime problem. When the partition is
no longer passive by the partition_manager it first notifies all
subscribers by sending unmanage notification (archiver_service). This
notification is not asynchronous. So it basically invokes a bunch of
callbacks. Then immediately it calls 'cluster::partition::stop' and when
the returned future is ready it deletes the partition. The FSM reacts
to unmanage notification by transitioning to 'stopping' state. It is
possible for the 'cluster::partition' to stop before the FSM. In this
case we have a race condition which may lead to triggering of the
assertion.

To avoid this every time we're starting/stopping the archiver we're
taking units from the _archiver_reset_mutex semaphore. The
'cluster::partition::stop' is trying to acquire units from this
semaphore first. This allows us to avoid race condition.

Note on state lifetimes. Boost MSM creates every state once during the
construction. Then it just tracks the current state internally. It never
creates or destroys states during transitions. Because of that state
transitions have to do some basic housekeeping. E.g. when FSM
transitions from 'starting' to 'active' state it has to release units
that it acquired from the '_archiver_reset_mutex'. Otherwise these units
will never be released.