archival: Remove 'sync' and introduce the alternative
Currently, the 'ntp_archiver' calls 'sync' method in every iteration to
avoid inconsistencies. It uses custom implementation that takes into
account the in-flight replication operation, current state of the raft
group (term, committed offset), and state of the manifest.

This commit changes this custom 'sync' implementation. Instead of
storing the in-flight operation we're just holding a lock and waiting
for the data in the log to be applied to the in-memory state of the STM.