storage: move ghost batch generation into log reader
Ghost batches are batches that don't contain data whose purpose is to
ensure:
- the caller (recovery_stm) sees a contiguous set of offsets
- the terms of the returned batches match with the original terms
  provided at append time, regardless of whether compactions have
  happened

These requirements are currently expected to be honored by the Raft
layer at recovery time, where it currently generates batches in between
batches read from storage. Since the storage layer is the current owner
of term metadata and compactions, the fact that these ghost batches are
generated by the Raft layer makes it trickier to evolve the underlying
storage implementation, since the guarantees would need to be tested /
validated at multiple layers.

So, to support development and testing of future log implementations,
this patch pulls out ghost batch generation into the storage layer as an
option to the log_reader.