admin: implement path parameter parser
This adds a utility method to parse/unescape path parameters in urls.

This was motivated by an issue where the admin API was using query
parameter unescaping to parse the username path parameter in ACL delete
and update requests, in which case the + sign was incorrectly replaced
with a ' '. This prevented admins from deleting users with a + in their
username.

(cherry picked from commit 27fe9c6c45936ba32278db07e6ce827c07ab9930)