admin: fix admin API path parameter parsing
This uses the new path parameter parsing method to correctly decode path
parameters in the admin API.

This was motivated by an issue where the admin API was using query
parameter unescaping to parse the username path parameter in ACL delete
and update requests, in which case the + sign was incorrectly replaced
with a ' '. This prevented admins from deleting users with a + in their
username.

(cherry picked from commit 02bb44eb9bf680d54b179e14cfb5daee4e2b143e)