test: rpk user delete of complex username
This adds a regression test to test that we can both create and delete a
user with a complex username (containing symbols that are considered to
be special symbols in the url syntax). This is tested through a ducktape
test using rpk to validate that the url encoding and decoding works well
throughout the interaction of rpk and the admin API.

This adds a regression test to test that we can both create and delete a
user with a complex username (containing symbols that are considered to
be special symbols in the url syntax). This is tested through a ducktape
test using rpk to validate that the url encoding and decoding works well
throughout the interaction of rpk and the admin API.

(cherry picked from commit c55f6030a2cce8f2dba845db4ad463b561c94748)