cloud_storage: Add 'applied_offset' to the manifest
The field contains an offset of the last applied command. It's advanced
by the 'archival_metadata_stm' every time the command is applied to the
in-memory state of the STM.

The commit contains a unit-test that checks that the field is serialized
correctly. The field is not added to the json manifest on purpose. The
json compatibility is tricky and it's OK for this field to be default
initialized after the manfiest reset.