cloud-tests: use a time-based calculation of load
For HighThroughputTest::test_decomm_and_add we used a fixed value of
10,000 messages to warm up the test, but with 4K messages and high
tiers this could result in a warmup time of < 1 second, which rounded
to zero, causing the test to time out immediately (in addition to being
an unsuitably small warmup even if it didn't time out).

Instead, use a time-based workload calculation, i.e., use the number
of messages we expect to be able to produce in a given time period:
this will naturally scale with tier size as expected throughput
increases.

Fixes #16470.