u/named_type: dropped wrapping curly braces
Dropped curly braces that all named types were wrapped with when printed
to string.

The curly braces that the string representation of `named_type` are
redundant and were often on purpose dropped by accessing the underlying
value while printing the `named_type`.

With this chance we are going to reduce the verbosity of log messages
and make them easier to read.

Signed-off-by: Michal Maslanka <michal@redpanda.com>