cluster: Add read_write_fence archival STM command
The command is supposed to be used as a concurrency control mechanism.
The user of the archival STM (ntp-archiver) should read the
applied_offset from the manifest before making any changes. Then, before
the segments are addded to the command builder (or any other changes)
the user should add `read_write_fence` command using the applied offset.
When the command is replicated the user will get an error if there was
any change to the in-memory state of the STM between the moment when the
applied_offset was acquired and the moment when the commands generated
by the user are applied. In other words, the read-write-fence guarantees
that the changes to the in-memory state are not made based on the stale
state. It guarantees that there is only one active writer.

The 'concurrent_modification_error' is returned if the property of the
read-write-fence is violated. The fence commmand is only applied to the
current batch. It can only abort commands which are added to the record
batch after it. If the read-write-fence command is added to the record
batch first it will abort the entire batch. If the command is added to
the record batch last it will not have any effect.