rptest: Combine SaslCredentials classes
We had two (identical) definitions of this class, but identical doesn't
mean "same type", so this doesn't type check.

Move the definition into a new redpanda_types module (since we have
to share a lot of basic types now between modules) and put an alias
in redpanda.py (where it originally lived) to avoid having to change
all the importers.