rptest: add assert_cluster_is_reusable
Add RedpandaServiceCloud.assert_cluster_is_reusable which is a more
comprehensive version of cluster_healthy(). In addition to checking
rpk's idea of health, it uses the CloudCluster _ensure_cluster health
as an additional check.

We run this at the end of cloud tests to ensure the cluster is still
healthy. This will catch more reasons the cluster may be damaged.

Issue #16331.