rptest: check brokers counts at end of test
In our end-of-test check for cloud tests, also check the broker counts.

We check the expected number of brokers from the tier definition against
the number of Redpanda pods reported by k8s as well as against the
/brokers endpoint from panda proxy, which is the number of brokers
the Redpanda cluster itself sees as part of the cluster.

This helps catch cases where tests may leave the cluster in not
reusable state because it has more or less brokers than the expected
number.