transform-sdk/js: Add Wasm JS VM
Since JavaScript cannot be compiled into Wasm, we take the (common)
approach of compiling a JavaScript runtime to Wasm and running the
interpreter in Wasm. We use the legandary QuickJS to run our JavaScript
within WebAssmebly. We're using a active fork of the original QuickJS
that fixes bugs and provides support for more modern ECMAScript
features. This QuickJS approach is similar to what the Bytecode
Alliance's Javy tool from Shopify does, except it is using the QuickJS
next-generation fork. I've contributed a few improvements to quickjs-ng
that will allow us to do 0 copy transforms into the JavaScript runtime
using typed arrays.

This commit adds a C++ wrapper to QuickJS to handle the reference
counting, and give a more modern C++ interface to the C APIs that
QuickJS provides. Not all functionality is exposed, but enough for the
functionality we need.

This also cargo cults some CMake from Redpanda proper to provide a CMake
based build system for this project, although this build system is for
internal use and we'll provide precompiled Wasm binaries for customers
to use.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>