raft: remove full_raft_configuration_recovery_pattern config property
The property was added as a workaround to a bug, but has since
(presumably) gone unused.

I have some upcoming changes to bring offset translation into the
storage layer, and this property raises a complication since it allows
Raft to start up with a mismatch between the storage log and Raft state.
Maintaining the property moving forward would make startup a bit more
difficult to reason about, and the property itself doesn't seem
particularly safe nonetheless.

This commit removes it outright.