s/readers_cache: added size based eviction to readers_cache
Added size based eviction to `storage::readers_cache`. Previously the
eviction was only time based which may lead to a problem when multiple
readers were created for the same NTP. Now with the size based eviction
the readers cache size is bounded and will not grow indefinitely.

Note on backward compatibility:

Previously the size of cache was unbounded. In this commit we change the
policy to keep up to 200 readers per ntp. This number is still large
enough not cause any issues as it would mean that there are 200
concurrent readers reading a single partition at the same time.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit a494de610eb2a44045ad74553eced3dc5dd30714)