r/stm_manager: release units immediately after finishing bckg apply
`raft::state_machine_manager` uses background apply fiber to
individually apply batches to state machines which are behind the main
apply fiber. When background apply fiber is active it reads and apply
batches up to current committed offset. When background apply fiber is
active it acquires the mutex. When mutex is acquired the main apply
fiber do not consider the stm as up to date.

The code was prone to very rare race condition as the background apply
was finished in one continuation but the units were release in
subsequent `finally` block. Normally this approach is harmless as the
semaphore is waited for and it will be signaled after the `finally`
fiber is executed. In `state_machine_manager` we only check if the
semaphore is available, this makes the solution vulnerable to timing.

Signed-off-by: Michal Maslanka <michal@redpanda.com>