cloud_storage: Improve stale_reader test
The test failed ocasionally when the test suite generated replacement
segments with the same size as original. The test replaced segments 1:1.
30 segments were replaced with 30 new randomly generated segments. This
caused the test to fail ocasionally when the size of the replacment
matced the size of the original segment.

This commit fixes that by replacing and merging segments. Every two old
segments are replaced with one new. The manifest allows replacement
segments to have any size if they are replacing more than one segment so
the test can no longer fail this way.