storage: make offset_translator use storage classes
The offset translator currently uses the storage::api to interact with
the storage layer. With it moving to the storage layer, it's more
convenient to have it instead use storage classes directly. For
instance, I intend on moving offset_translator into the disk_log_impl,
where storage_api doesn't exist.

This commit makes the translator use the kvstore and storage_resources
directly, and introduces a new constructor that avoids storage_api.