storage: add interface to tell if log is brand new
This check is already done by Raft to tell whether it should reset
offset translator state. With the offset translator moving into the
storage layer, it makes sense for this check to be done directly in
storage.

Note that Raft also previously checked the value of
`_last_snapshot_index`. This does not seem required, as the last
snapshot index is only hydrated in hydrate_snapshot() during do_start()
after Raft checks whether the log is new.