storage: add offset translator option to force truncate
Raft currently has the capability of force truncating the offset
translator state when there is a gap in the log, using
offset_translator::prefix_truncate_reset(). It does this in the
unexpected event that the kvstore doesn't match with the contents of the
log.

Rather than having this special case of offset-translator truncation,
this commit introduces an truncation option to allow for this forced
truncation, expecting the caller to supply an offset delta to use.

This commit replaces the implementation so that prefix_truncate_reset()
is just a call to the more expected prefix_truncate(). This will make it
easier to have storage::disk_log_impl own the offset translator, without
having to worry about special Raft-specific usage.