storage: instantiate offset_translator in disk_log_impl
This commit instantiates a currently unused instance of a
raft::offset_translator in the disk_log_impl. This will be required as
offset translation is pulled into the storage layer.

This change alone, while mostly non-functional, at least allows
subsequent commits to make further changes that depend on the new
storage-layer translator without switching over to using it.