rpk/transform/logs: add initial command skeleton
Add the initial boilerplate for the command, as well as the flags. This
includes adding the timequery parsing flags. The parsing code here is
very similar to the timequery parsing flags in `rpk topic consume`,
except they are simplified a little bit.

Tests are added to ensure parsing is correct.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>