cloud_storage/anomalies_detector: add max_num_segments limit
max_num_segments limits the total number of segment_meta that will be
checked in a run. To check a segment meta the code checks the existance
of its segment in remote storage. This done with a HEAD request.

the limit is counted from most recent to oldest, so that a depth of 1 will
only cause the newest segment to be checked.

an appropriate low number will limit the scrub to only the stm manifest
section.

we could limit by
- num segments
- offset
- total time

a limit by num segments correlates directly with the speed
of the operation while being deterministic.

offset could map better with what the user wants, but the current implementation
of reverse iteration in segment_meta_cstore makes it a bit involved to implement.

total time it's a strong guarantee but the final result depends on the load of the system.

This commit implements a limit by num of segments with the parameter
max_num_segments. The implementation of this mode it's easier to reason about,
and other PRs can implement the other two modes.