c/leader_balancer: replaced muted group index with roaring bitmap
The muted group index was using a `absl::flat_has_set` the use of flat
data structure lead to oversized allocation when the number of
muted/skipped raft group was large. Replaced a hash set with roaring
bitmap. The roaring bitmap is efficient and compressed leading to better
performance and smaller allocation.

Fixes: #12005

Signed-off-by: Michal Maslanka <michal@redpanda.com>