kafka/test: create topics response
This adds a regression test to ensure that the configs returned by
describe configs and the configs returned by create topics match.

This updates/removes an earlier test which only checked that a single
specific config was present.

NOTE: the old test I am updating would fail with the current code. That
is because the new code returns a different source for the
cleanup.policy config, namely DYNAMIC_TOPIC_CONFIG (source=1) instead of
DEFAULT_CONFIG (source=5). Apache Kafka (and our DescribeConfigs
endpoint) both return source=5 in this case, so I believe this is a
bugfix rather than a regression.