kafka: fix config source for cleanup policy
Currently Redpanda returns source=1 (topic-specific overwrite) for the
cleanup.policy config. This is different to Apache Kafka, which returns
source=5 (default config), even though the config value they return
(delete) is the same, so I believe this behaviour is a bug and this
commit fixes it by using the hide_default_override helper field.