rpk/transform: implement logs command
Implement the logs command for transforms.

This command is very similar in general structure to `rpk topic
consume`, as they are fundamentally doing similar things.

The differences is that we only care about a single partition, so we
lookup the partition the transform lives in and then consume all logs
within the bounds of our query and print them.

There is some awkwardness as we need to buffer log messages to properly
emit newlines (otherwise we could get partial log lines split on
multiple lines of output, and we split warning logs and info logs on
different lines). Otherwise we add timestamps and levels to the output
when one does `--format=wide`.

Lastly, the `--tail` flag is hard to implement using franz-go. One could
make a new client everytime and incrementally read batches of records in
reverse, but that seems complex, so we just instead buffer everything
then dump the logs at the end.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>