tests/cluster_config_test test_aliasing fix rpk config version
when setting a new config the response is the new config version. before
reading back we need to wait for this version to propagate to the whole
cluster.

extracts from the result of rpk.cluster_config_set the new config
version and wait for propagation, before trying to read it again.

redpanda.set_cluster_config internally does the same wait