rptest: Fix cloud_retention_test
Fix retention test. Previously, the throughput parameter was passed to
the parameter that limited max number of messages. Because of that the
test was only able to consume fixed number of messages in CDT. This
commit fixes this and improves few other aspects (error handling and
logging).

(cherry picked from commit 042e4c2a8754106663bb02579715023117e35481)