wasm/wasi: expose all output topic names
Previously the output topic was written as `REDPANDA_OUTPUT_TOPIC`, now
that there can be multiple we'll write each as
`REDPANDA_OUTPUT_TOPIC_%d` with the index as the key. This is
technically a breaking change to remove the `REDPANDA_OUTPUT_TOPIC` key
for the first output topic, but since this isn't documented anywhere and
even I haven't used it it's going hurt anything here.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>