cst/inv: get latest report metadata
A new method is added which performs the following actions:

* scan the bucket for inventory reports
* find the latest report which is ready for download
* download the manifest for report and extract paths to report files

Finally metadata for the report is returned which contains the path to
manifest, paths to actual reports and the date-time as an ISO-8601
formatted string.

This metadata can be used by the caller to determine if it wants the
report (i.e. the date-time is newer than last consumed report by caller)
and to actually download the report files.