c/contiguous_range_map: introduced contiguous_range_map
Contiguous range map is an associative sorted container backed by
chunked_vector designed to efficiently store objects indexed with
contiguous and limited range of integers. The container provides a
wrapper around basic chunked vector that reassembles API of a standard map.

The wrapper allows random inserts of elements to the map event tho this is
not supported by standard vector.

Underlying container is resized every time its size is not sufficient to
account for emplaced key.

The contiguous_range_map tolerates gap in the range of keys however
number and size of gaps is proportional to performance penalty hit when
incrementing or decrementing iterators.

Signed-off-by: Michal Maslanka <michal@redpanda.com>