c/leader_balancer: remove unneeded code
Removed piece assumes that partition.replicas contains only followers
and therefore contains assignment.to, but not assignment.from (under
this assumption it changes assignment.to to assignment.from). But in
reality partition.replicas contains all replicas, and therefore both
assignment.to and assignment.from. As other code doesn't make any
assumptions about presence or absence of the leader shard from the list
of replicas in the index, we can simply use the original list of
replicas.