utils/mutex: remove unnamed constructor
Certain structures like semaphore and mutex can be "named" which is
useful to identify a particular structure in log messages, exceptions
or just about anything you can imagine.

When we added support for named mutexes, we did not give names to all
existing mutexes, perhaps because there are a lot of them: the
unnamed version of the constructor was left and remove it was
left as a TODO.

This change crosses that TODO off the list and removes the unnamed
constructor (the prior change in this series named all existing
mutexes that weren't already named).

Issue #5489.