transform-sdk/integration: initial suite
Add an identity transform integration test for all the different SDKs we
will support.

This is a generic runner that will just take Wasm files as input. We
will use this to add tests for each of the SDKs to ensure they work with
the broker. These tests are less expensive than running each one within
the broker's test suite, so we have this seperate smaller test suite
that ensures the SDKs are doing the right thing, since the broker's
tests are more about validating the broker is doing the right thing.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>