storage: add translator type options to log builder
Some tests use the translator in the log builder to do translation, and
will need to be able to supply a reasonable set of batch types to
translate.

Since the on-disk storage of the offset translator is in the kvstore
keyed by group_id, this also adds an optional group_id to the builder.