storage: introduce metrics for space management
The goal of space management related metrics are to provide
observability and insight into what space management is doing, if it is
working correctly, and if not able to reclaim data, what is holding
things up.

This commit introduces the the public/internal metrics registration, and
seeds the metrics with the "target_size" metric exposing the target size
of the data disk, usage beyond which should trigger data reclaim.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>