rptest: correct offset in transform logging test
Offsets are 0 indexed, so the last record in the transform should be n -
1.

NOTE: It's still possible for this test to be flaky if all the input
partitions end up on the same core, as all of the partitions could
produce logs at once, overshooting the log buffer. Maybe that is
unlikely enough it will never happen.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>
(cherry picked from commit bd6ed7fa7c5ad908fa29d584bf6171f6d8e0dec7)