cloud_storage/topic_manifest: DOM ser/des via enum keys
The functions are not wired in yet.

Major changes: Introduced DOM serialization/deserialization, and an adapter
to be able to use enums as keys while reading/writing k:v pairs in the
json representation.

The ability to use enums allows to place a static check that fields are
in sync with topic_properties.

To add a new field, the user needs to add a new enum fields value, add a
line in do_serialization and a line in do_update.

The deserialization uses the information of which fields are
introduced in every topic_properties::redpanda_serde_version to
correctly handle tristate properties.