archival: Add async_data_uploader class
The async_data_uploader can be used to provide the
ss::input_stream<char> that cover certain offset range and generate
metadata for it. The class makes a storage::reader and uses it to go
through the offset range while serializing everything into disk format.
The reader is bypassing both batch cache and readers cache to avoid
interference with kafka fetch requests and avoid cache pollution.

The uploader can upload offset ranges when inclusive range is provided.
It can also find the upload that matches search predicate that includes
start offset, desired size and smallest acceptable upload size. In this
case the upload always ends on an index boundary so the next search will
start on the index boundary as well. This eliminates the necessity to
scan the segment in order to find the boundary.

The sizing is not precise and can overshoot by up to 32KiB.