archival: Add async_data_uploader tests
All tests are generating segments with data and different randomized
test cases. The size limited mode is only tested with non-compacted
segments because the compacted segments are only reuploaded.

The use of rpfixture in this case is justified because every fixture
test generates data once and then uses it to run hundereds of test
cases.