c/hm_backend: do not request self node report from the controller
Health monitor backend on every node contains a health reports from all
cluster nodes. The health reports are collected on the current
controller node and then cached on the others. Previously a node asked
controller to send it back a report that was collected by the controller
leading to unnecessary traffic and serialization related computation.

Change the logic so that the node collect the health report on its own
while query the controller for other cluster member reports.

Signed-off-by: Michal Maslanka <michal@redpanda.com>