c/shard_placement_table: get rid of shard-local targets
Previously we stored the target shard for the ntp on all shards with
state related to that ntp. This way during a cross-shard transfer the
source shard knows immediately where to transfer state to. The drawback
is that it is hard to synchronize state updates coming from
different controller_backend reconciliation fibers and from updating
targets. Introduce a simple assignment marker instead that shows if the
partition should be present or not on the current shard (to get the
destination shard we have to query the node-wide map on shard 0)