container: add interval_set
This adds an interval_set class, similar to interval_map with a couple
key differences:
- it is a set type, and only tracks the intervals rather than an
  additional value type
- it supports inserting overlapping intervals by coalescing them

This will be a used in an upcoming change to track regions within a file
to skip over, which will be key to a append-only log implementation that
truncates by appending metadata to the log.

The structure and some of the implementation of the interval_set is
copied directly from interval_map, though the insert() method is
significantly different, given the above behavioral differences.