rptest: fix _lifecycle_test naming
Ducktape uses method naming to decide what is a test and our
ManyTopicsTest had a _lifecycle_test method which is a helper called
by tests and not a standalone test itself, but DT will call it as
a standalone test causing a CI failure.

Rename it to _lifecycle_test_impl to avoid this.

Fixes #17040.