schema_registry: Override soft-delete checks for consume_to_store
It's possible that a soft-delete can be compacted away in a different
segment to the hard delete. After compaction the topic is still correct
and the commands should always be applied to the store.

Override the soft-delete check if the command comes from the topic
rather than the REST API.

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit bf1c314b51a25cfd7d125258495652db53569f9e)