rpk: take into account auth kind to clear profiles
If we have profile A from auth SSO, we cannot keep profile A if we
subsequently log in with client credentials. Even though the
organizations are the same, the auth is different, and we need a new
profile tied to the new auth.

(cherry picked from commit 9c6ec387df6613d48bca7ea1089b78ee370e118e)