r/stm_manager: fixed not applying snapshot for new stms
State machine manager manages the aggregated Raft snapshot for all the
state machines created based on one Raft instance. The managed snapshot
is a map containing individual snapshot data for each state machine.
When a new STM is created while the managed snapshot was already taken
the `state_machine_base::apply_raft_snapshot` should still be called
even if the snapshot doesn't exists in the managed snapshot map.
This way an STM will be informed that the log doesn't start from 0.

Previously when snapshot was not present in the managed snapshot map we
skipped calling `apply_snapshot` on the STM and didn't advance it's
`_next` offset which lead to stuck background apply fiber loop.

Signed-off-by: Michal Maslanka <michal@redpanda.com>