raft: refactor snapshot hydration
An upcoming change will rewrite the initial snapshot hydration at
startup to use the new log::start() interface. To that end, this pulls
some logic from snapshot hydration into some reusable methods.

Note, there is a mechanical difference in behavior that in practice will
result in the same behavior:

Previously, we always truncated the offset translator after adding the
latest config to the config manager, and then conditionally truncated
the log (roughly, we wouldn't truncate if this is the controller
partition).

With this change, the truncation of the offset translator will only
happen if the log is also going to be truncated. In practice, this
behavioral change doesn't matter because not truncating the log implies
this is the controller partition, which doesn't translate offsets
anyway (its translated batch types is empty).